Implement delete tagg

+++ app/src/main/java/com/example/instatagg/data/dbTaggs/TaggsDBDataSource.kt
@@ -17,7 +17,7 @@ class TaggsDBDataSource(
         taggsDao.changeTaggName(name, newTagg)
     }
 
-    fun delTagg(id: Int) {
+    fun delTagg(id: Long) {
         taggsDao.delTagg(id)
     }
